// src/utils/enums.ts
export enum userAssetTabs {
  COLLECTED = 'collected',
  CREATED = 'created',
  CREATED_COLLECTIONS = 'createdCollections',
  FAVORITED = 'favorited',
  ACTIVITY = 'activity',
  OFFERS_RECEIVED = 'offersReceived',
  OFFERS_MADE = 'offersMade',
  LISTINGS_ACTIVE = 'listingsActive',
  LISTINGS_INACTIVE = 'listingsInactive',
}
export enum userSortOptions {
  PRICE = 'price',
  LISTING_DATE = 'listingDate',
}
export enum userSortOrderOptions {
  ASC = 'ASC',
  DESC = 'DESC',
}
export enum userFilterStatuses {
  HAS_OFFERS = 'hasOffers',
  NEW = 'new',
  BUY_NOW = 'buyNow',
  ON_AUCTION = 'onAuction',
}
export enum priceFilterOptions {
  USD = 'USD',
  SOL = 'SOL',
  ETH = 'ETH',
}